+++ lib/aliasDocVariables_.d.ts
@@ -1,0 +1,8 @@
+import { DocumentNode } from 'graphql';
+import { Variables } from './types';
+interface CachedExecParams {
+    document: DocumentNode;
+    variables: Variables;
+}
+declare const aliasDocVariables_: (execParams: CachedExecParams, aliasIdx: number, baseVariables: Variables) => DocumentNode;
+export default aliasDocVariables_;