+++ lib/renameResponseTypenames_.d.ts
@@ -1,3 +1,3 @@
-import { EndpointExecutionResult } from './types';
-declare const renameResponseTypenames: (response: EndpointExecutionResult, prefix: string) => void;
+import { AliasMap, EndpointResponseData } from './types';
+declare const renameResponseTypenames: (response: EndpointResponseData, prefix: string, aliasMaps: AliasMap[]) => void;
 export default renameResponseTypenames;